Refactor OSProcess object, add tests
