--- oshi-core/src/main/java/oshi/software/os/windows/WindowsProcess.java
@@ -1,34 +1,0 @@
-/**
- * Oshi (https://github.com/dblock/oshi)
- *
- * Copyright (c) 2010 - 2017 The Oshi Project Team
- *
- * All rights reserved. This program and the accompanying materials
- * are made available under the terms of the Eclipse Public License v1.0
- * which accompanies this distribution, and is available at
- * http://www.eclipse.org/legal/epl-v10.html
- *
- * Maintainers:
- * dblock[at]dblock[dot]org
- * widdis[at]gmail[dot]com
- * enrico.bianchi[at]gmail[dot]com
- *
- * Contributors:
- * https://github.com/dblock/oshi/graphs/contributors
- */
-package oshi.software.os.windows;
-
-import oshi.software.common.AbstractProcess;
-
-/**
- * A process is an instance of a computer program that is being executed. It
- * contains the program code and its current activity. Depending on the
- * operating system (OS), a process may be made up of multiple threads of
- * execution that execute instructions concurrently.
- *
- * @author widdis[at]gmail[dot]com
- */
-public class WindowsProcess extends AbstractProcess {
-
-    private static final long serialVersionUID = 1L;
-}
