+++ oshi-json/src/test/java/oshi/json/software/os/OperatingSystemTest.java
@@ -54,6 +54,11 @@ public class OperatingSystemTest {
         OSProcess proc = os.getProcess(os.getProcessId());
         assertTrue(proc.getName().length() > 0);
         assertTrue(proc.getPath().length() > 0);
+        assertTrue(proc.getCommandLine().length() > 0);
+        assertTrue(proc.getUser().length() > 0);
+        assertTrue(proc.getUserID().length() > 0);
+        assertNotNull(proc.getGroup());
+        assertNotNull(proc.getGroupID());
         assertNotNull(proc.getState());
         assertEquals(proc.getProcessID(), os.getProcessId());
         assertTrue(proc.getParentProcessID() > 0);
@@ -67,5 +72,58 @@ public class OperatingSystemTest {
         assertTrue(proc.getStartTime() >= 0);
         assertTrue(proc.getBytesRead() >= 0);
         assertTrue(proc.getBytesWritten() >= 0);
+    }
+
+    /**
+     * Test OSProcess setters and getters
+     */
+    @Test
+    public void testOSProcessSetters() {
+        SystemInfo si = new SystemInfo();
+        OperatingSystem os = si.getOperatingSystem();
+        OSProcess oldProcess = os.getProcess(os.getProcessId());
+
+        OSProcess newProcess = new OSProcess();
+        newProcess.setName(oldProcess.getName());
+        newProcess.setPath(oldProcess.getPath());
+        newProcess.setCommandLine(oldProcess.getCommandLine());
+        newProcess.setUser(oldProcess.getUser());
+        newProcess.setUserID(oldProcess.getUserID());
+        newProcess.setGroup(oldProcess.getGroup());
+        newProcess.setGroupID(oldProcess.getGroupID());
+        newProcess.setState(oldProcess.getState());
+        newProcess.setProcessID(oldProcess.getProcessID());
+        newProcess.setParentProcessID(oldProcess.getParentProcessID());
+        newProcess.setThreadCount(oldProcess.getThreadCount());
+        newProcess.setPriority(oldProcess.getPriority());
+        newProcess.setVirtualSize(oldProcess.getVirtualSize());
+        newProcess.setResidentSetSize(oldProcess.getResidentSetSize());
+        newProcess.setKernelTime(oldProcess.getKernelTime());
+        newProcess.setUserTime(oldProcess.getUserTime());
+        newProcess.setUpTime(oldProcess.getUpTime());
+        newProcess.setStartTime(oldProcess.getStartTime());
+        newProcess.setBytesRead(oldProcess.getBytesRead());
+        newProcess.setBytesWritten(oldProcess.getBytesWritten());
+
+        assertEquals(oldProcess.getName(), newProcess.getName());
+        assertEquals(oldProcess.getPath(), newProcess.getPath());
+        assertEquals(oldProcess.getCommandLine(), newProcess.getCommandLine());
+        assertEquals(oldProcess.getUser(), newProcess.getUser());
+        assertEquals(oldProcess.getUserID(), newProcess.getUserID());
+        assertEquals(oldProcess.getGroup(), newProcess.getGroup());
+        assertEquals(oldProcess.getGroupID(), newProcess.getGroupID());
+        assertEquals(oldProcess.getState(), newProcess.getState());
+        assertEquals(oldProcess.getProcessID(), newProcess.getProcessID());
+        assertEquals(oldProcess.getParentProcessID(), newProcess.getParentProcessID());
+        assertEquals(oldProcess.getThreadCount(), newProcess.getThreadCount());
+        assertEquals(oldProcess.getPriority(), newProcess.getPriority());
+        assertEquals(oldProcess.getVirtualSize(), newProcess.getVirtualSize());
+        assertEquals(oldProcess.getResidentSetSize(), newProcess.getResidentSetSize());
+        assertEquals(oldProcess.getKernelTime(), newProcess.getKernelTime());
+        assertEquals(oldProcess.getUserTime(), newProcess.getUserTime());
+        assertEquals(oldProcess.getUpTime(), newProcess.getUpTime());
+        assertEquals(oldProcess.getStartTime(), newProcess.getStartTime());
+        assertEquals(oldProcess.getBytesRead(), newProcess.getBytesRead());
+        assertEquals(oldProcess.getBytesWritten(), newProcess.getBytesWritten());
     }
 }
