+++ oshi-json/src/test/resources/oshi.json.properties
@@ -76,6 +76,11 @@
 #  operatingSystem.processes                             = false
 #    operatingSystem.processes.name                      = false
 #    operatingSystem.processes.path                      = false
+#    operatingSystem.processes.commandLine               = false
+#    operatingSystem.processes.user                      = false
+#    operatingSystem.processes.userID                    = false
+#    operatingSystem.processes.group                     = false
+#    operatingSystem.processes.groupID                   = false
 #    operatingSystem.processes.state                     = false
 #    operatingSystem.processes.processID                 = false
 #    operatingSystem.processes.parentProcessID           = false
